# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Unlock Wealth Resort API",
        "description": "API documentation for Unlock Wealth Resort",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/users": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Get a list of users",
                "description": "Returns a paginated list of all users.",
                "operationId": "9d383669d3cdd73072d5665613a02f16",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Users retrieved successfully"
                                        },
                                        "data": {
                                            "properties": {
                                                "items": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/components/schemas/UserResource"
                                                    }
                                                },
                                                "pagination": {
                                                    "properties": {
                                                        "total": {
                                                            "type": "integer",
                                                            "example": 50
                                                        },
                                                        "per_page": {
                                                            "type": "integer",
                                                            "example": 15
                                                        },
                                                        "current_page": {
                                                            "type": "integer",
                                                            "example": 1
                                                        },
                                                        "last_page": {
                                                            "type": "integer",
                                                            "example": 4
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Create a new user",
                "description": "Creates a new user account.",
                "operationId": "64a7fb6b692c14ebf6c715ef0b29be3c",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "fullName",
                                    "nickname",
                                    "email",
                                    "password",
                                    "address",
                                    "profile",
                                    "dateOfBirth",
                                    "placeOfBirth",
                                    "weight",
                                    "telegramPh",
                                    "whatsappPh"
                                ],
                                "properties": {
                                    "fullName": {
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "nickName": {
                                        "type": "string",
                                        "example": "John"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "johndoe@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "Password123"
                                    },
                                    "address": {
                                        "type": "string",
                                        "example": "Rose Mary Street"
                                    },
                                    "profile": {
                                        "type": "string",
                                        "example": "helloworld.jpg"
                                    },
                                    "dateOfBirth": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "01-01-1000"
                                    },
                                    "placeOfBirth": {
                                        "type": "string",
                                        "example": "Yangon"
                                    },
                                    "weight": {
                                        "type": "integer",
                                        "example": "56"
                                    },
                                    "telegramPh": {
                                        "type": "string",
                                        "example": "09xxxxxxxxx"
                                    },
                                    "whatsappPh": {
                                        "type": "string",
                                        "example": "09xxxxxxxxx"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "User created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserResource"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "500": {
                        "description": "User creation failed"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/users/{id}": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Get a single user",
                "description": "Returns the details of a specific user by their ID.",
                "operationId": "765733b743508faab1026cd873f66b07",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the user",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User Fetched Successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "User not found"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Users"
                ],
                "summary": "Update an existing user",
                "description": "Updates the details of an existing user.",
                "operationId": "e2d4a34e8d1361706359213c183212dc",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the user to update",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "fullName",
                                    "nickname",
                                    "email",
                                    "address",
                                    "dateOfBirth",
                                    "placeOfBirth",
                                    "weight"
                                ],
                                "properties": {
                                    "fullName": {
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "nickName": {
                                        "type": "string",
                                        "example": "John"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "johndoe@example.com"
                                    },
                                    "address": {
                                        "type": "string",
                                        "example": "Rose Mary Street"
                                    },
                                    "dateOfBirth": {
                                        "type": "string",
                                        "example": "01-01-1000"
                                    },
                                    "placeOfBirth": {
                                        "type": "string",
                                        "example": "Yangon"
                                    },
                                    "weight": {
                                        "type": "integer",
                                        "example": "56"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User Updated Successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserResource"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "404": {
                        "description": "User not found"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "UserResource": {
                "title": "User Resource",
                "description": "User model representation",
                "properties": {
                    "id": {
                        "description": "User ID",
                        "type": "integer",
                        "example": 1
                    },
                    "Fullname": {
                        "description": "Fullname of the user",
                        "type": "string",
                        "example": "John Doe"
                    },
                    "Nickname": {
                        "description": "Nickname of the user",
                        "type": "string",
                        "example": "John"
                    },
                    "email": {
                        "description": "Email of the user",
                        "type": "string",
                        "format": "email",
                        "example": "johndoe@example.com"
                    },
                    "profileUrl": {
                        "description": "Profile url of the user",
                        "type": "string",
                        "example": "example.jpg"
                    },
                    "profilePublicId": {
                        "description": "Profile Id of the user",
                        "type": "string",
                        "example": "1"
                    },
                    "address": {
                        "description": "Profile Id of the user",
                        "type": "string",
                        "example": "Rose Mary Street"
                    },
                    "TelegramPh": {
                        "description": "Telegram number of the user",
                        "type": "string",
                        "example": "09xxxxxxx"
                    },
                    "WhatsappPh": {
                        "description": "Whatsapp number of the user",
                        "type": "string",
                        "example": "09xxxxxxx"
                    },
                    "dateOfBirth": {
                        "description": "Birthday of the user",
                        "type": "string",
                        "example": "10-01-19xx"
                    },
                    "placeOfBirth": {
                        "description": "Birth place of the user",
                        "type": "string",
                        "example": "Yangon"
                    },
                    "weight": {
                        "description": "weight of the user",
                        "type": "integer",
                        "example": "56"
                    },
                    "dailyRoutine": {
                        "description": "Daily routine of the user",
                        "type": "text",
                        "example": "Lorem ipsum dolor sit amet consectetur adipisicing elit. Perferendis beatae reiciendis quia iure id debitis, accusantium nisi et fugit alias temporibus aspernatur pariatur obcaecati sint. Ullam quae nostrum debitis eum"
                    },
                    "specialRequest": {
                        "description": "Special request of the user",
                        "type": "text",
                        "example": "Lorem"
                    },
                    "isVerified": {
                        "description": "Email verified of the user",
                        "type": "boolean",
                        "example": "true"
                    },
                    "isPremium": {
                        "description": "Premium of the user",
                        "type": "boolean",
                        "example": "false"
                    },
                    "startDate": {
                        "description": "Premium start date of the user",
                        "type": "string",
                        "format": "date",
                        "example": "01-01-20xx"
                    },
                    "endDate": {
                        "description": "Premium end date of the user",
                        "type": "string",
                        "format": "date",
                        "example": "01-01-20xx"
                    },
                    "createdAt": {
                        "description": "Creation timestamp",
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-10-28T12:00:00.000000Z"
                    },
                    "updatedAt": {
                        "description": "Last update timestamp",
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-10-28T12:00:00.000000Z"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}